+++ spec/integrations/broadcastable_subscriptions_spec.rb
@@ -42,9 +42,8 @@ RSpec.describe "broadcastable subscriptions" do
 
       stream_name = subject.streams.first
 
-      # update request context and channelId
+      # update request context
       request.connection_identifiers = identifiers.merge(current_user: "alice").to_json
-      request.identifier = channel_identifier.merge(channelId: rand(1000).to_s).to_json
 
       response = handler.handle(:command, request)
       expect(response).to be_success
@@ -52,7 +51,6 @@ RSpec.describe "broadcastable subscriptions" do
 
       # now update the query param
       request.data = data.merge(variables: {id: "b"}).to_json
-      request.identifier = channel_identifier.merge(channelId: rand(1000).to_s).to_json
 
       response = handler.handle(:command, request)
       expect(response).to be_success
@@ -114,9 +112,8 @@ RSpec.describe "broadcastable subscriptions" do
 
       request_2 = request.dup
 
-      # update request context and channelId
+      # update request context
       request_2.connection_identifiers = identifiers.merge(current_user: "alice").to_json
-      request_2.identifier = channel_identifier.merge(channelId: rand(1000).to_s).to_json
 
       response_2 = handler.handle(:command, request_2)
 
@@ -155,47 +152,6 @@ RSpec.describe "broadcastable subscriptions" do
 
       schema.subscriptions.trigger(:post_updated, {id: "a"}, POSTS.first)
       expect(AnyCable.broadcast_adapter).to have_received(:broadcast).twice
-    end
-
-    context "with similar ChannelId and not using ID from client" do
-      let(:config) { GraphQL::AnyCable.config }
-
-      around do |ex|
-        config.use_client_provided_uniq_id.tap do |was_val|
-          config.use_client_provided_uniq_id = false
-          ex.run
-          config.use_client_provided_uniq_id = was_val
-        end
-      end
-
-      specify "creates an entry for each subscription" do
-        # first, subscribe to obtain the connection state
-        subscribe_response = handler.handle(:command, request)
-        expect(subscribe_response).to be_success
-  
-        expect(redis.keys("graphql-subscription:*").size).to eq(1)
-        expect(redis.keys("graphql-subscriptions:*").size).to eq(1)
-  
-        # update request context
-        request.connection_identifiers = identifiers.merge(current_user: "alice").to_json
-  
-        response = handler.handle(:command, request)
-  
-        expect(redis.keys("graphql-subscription:*").size).to eq(2)
-        expect(redis.keys("graphql-subscriptions:*").size).to eq(1)
-  
-        istate = response.istate
-  
-        request.command = "unsubscribe"
-        request.data = ""
-        request.istate = istate
-  
-        response = handler.handle(:command, request)
-        expect(response).to be_success
-  
-        expect(redis.keys("graphql-subscription:*").size).to eq(1)
-        expect(redis.keys("graphql-subscriptions:*").size).to eq(1)
-      end
     end
 
     context "without subscription" do